renamed /user route to /createUser route to reflect how we desgined it in swagger
